+++ sources/packages/api/src/app.aest.module.ts
@@ -55,10 +55,12 @@ import {
   ApplicationExceptionAESTController,
   ScholasticStandingAESTController,
   ScholasticStandingControllerService,
+  InstitutionUserAESTController,
 } from "./route-controllers";
 import { AuthModule } from "./auth/auth.module";
 import { LoggerModule } from "./logger/logger.module";
 import { ApplicationControllerService } from "./route-controllers/application/application.controller.service";
+import { InstitutionUserControllerService } from "./route-controllers/institution-user/institution-user.controller.service";
 
 @Module({
   imports: [LoggerModule, DatabaseModule, AuthModule],
@@ -67,6 +69,7 @@ import { ApplicationControllerService } from "./route-controllers/application/ap
     DesignationAgreementAESTController,
     ApplicationAESTController,
     InstitutionAESTController,
+    InstitutionUserAESTController,
     AssessmentAESTController,
     StudentAESTController,
     StudentAppealAESTController,
@@ -118,6 +121,7 @@ import { ApplicationControllerService } from "./route-controllers/application/ap
     ScholasticStandingControllerService,
     SINValidationService,
     InstitutionUserAuthService,
+    InstitutionUserControllerService,
   ],
 })
 export class AppAESTModule {}
