+++ sources/packages/api/src/route-controllers/institution-locations/models/institution-location.dto.ts
@@ -120,3 +120,9 @@ export enum DesignationStatus {
    */
   NotDesignated = "Not designated",
 }
+
+export class InstitutionLocationsAPIOutDTO {
+  id: number;
+  name: string;
+  address: AddressAPIOutDTO;
+}
