+++ sources/packages/api/src/utilities/index.ts
@@ -16,7 +16,6 @@ export * from "./disbursements-utils";
 export * from "./datatable-database-mapping-utils";
 export * from "./note-utils";
 export * from "./datatable-config";
-export * from "./institution-utils";
 export * from "./database";
 export * from "./pagination-utils";
 export * from "./address-utils";
