+++ sources/packages/web/src/components/institutions/modals/EditInstitutionUserModal.vue
@@ -59,6 +59,7 @@ import {
   INSTITUTION_USER_ALREADY_EXISTS,
   LEGAL_SIGNING_AUTHORITY_EXIST,
 } from "@/constants";
+import { InstitutionUserService } from "@/services/InstitutionUserService";
 
 const submitKnownErrors = [
   INSTITUTION_USER_ALREADY_EXISTS,
@@ -116,7 +117,7 @@ export default {
       userInfo.value = params;
       // Get the user permissions.
       const userDetails =
-        await InstitutionService.shared.getInstitutionLocationUserDetails(
+        await InstitutionUserService.shared.getInstitutionUserByUserName(
           userInfo.value.userName,
         );
       // A user is considered an admin if any authorization has a userType defined as admin.
@@ -161,7 +162,7 @@ export default {
         processing.value = true;
         const userManagementModel = institutionUserManagement.value
           .formModel as UserManagementModel;
-        await InstitutionService.shared.updateInstitutionUserWithAuth(
+        await InstitutionUserService.shared.updateInstitutionUserWithAuth(
           userInfo.value.userName,
           userManagementModel.isAdmin,
           userManagementModel.isLegalSigningAuthority,
