+++ sources/packages/web/src/helpers/helpers/uri.ts
@@ -1,4 +1,9 @@
-import { DataTableSortOrder, FieldSortOrder, PaginationParams } from "@/types";
+import {
+  DataTableSortOrder,
+  FieldSortOrder,
+  PaginationOptions,
+  PaginationParams,
+} from "@/types";
 
 /**
  * helper to append pagination sort and order to the url
@@ -35,7 +40,41 @@ export const addPaginationOptions = (
   url: string,
   page: number,
   pageLimit: number,
-  joinString: string,
+  joinString = "",
 ): string => {
   return `${url}${joinString}${PaginationParams.Page}=${page}&${PaginationParams.PageLimit}=${pageLimit}`;
+};
+
+/**
+ * Builds the query string parameters for pagination.
+ * @param paginationOptions pagination options.
+ * @returns the URL query string in a format like parameter1=value1&parameter2=value2.
+ */
+export const getPaginationQueryString = (
+  paginationOptions: PaginationOptions,
+) => {
+  const parameters: string[] = [];
+  // Paginations parameters.
+  parameters.push(`${PaginationParams.Page}=${paginationOptions.page}`);
+  parameters.push(
+    `${PaginationParams.PageLimit}=${paginationOptions.pageLimit}`,
+  );
+  // Sort parameters.
+  if (paginationOptions.sortField && paginationOptions.sortOrder) {
+    const sortDBOrder =
+      paginationOptions.sortOrder === DataTableSortOrder.DESC
+        ? FieldSortOrder.DESC
+        : FieldSortOrder.ASC;
+    parameters.push(
+      `${PaginationParams.SortField}=${paginationOptions.sortField}`,
+    );
+    parameters.push(`${PaginationParams.SortOrder}=${sortDBOrder}`);
+  }
+  // Search criteria.
+  if (paginationOptions.searchCriteria) {
+    parameters.push(
+      `${PaginationParams.SearchCriteria}=${paginationOptions.searchCriteria}`,
+    );
+  }
+  return parameters.join("&");
 };
