+++ sources/packages/web/src/services/InstitutionService.ts
@@ -4,22 +4,14 @@ import {
   InstitutionUserTypes,
   LocationAuthorization,
   LocationUserAccess,
-  PaginationParams,
 } from "@/types";
 import {
-  InstitutionLocationsDetails,
-  InstitutionUserAuthDetails,
-  InstitutionUserViewModel,
-  InstitutionUserRoleLocation,
-  InstitutionUserWithUserType,
   OptionItemDto,
   DataTableSortOrder,
-  InstitutionUserAndCountForDataTable,
   PaginatedResults,
   PaginationOptions,
 } from "../types";
 import ApiClient from "./http/ApiClient";
-import { AuthService } from "./AuthService";
 import {
   InstitutionLocationFormAPIInDTO,
   InstitutionLocationFormAPIOutDTO,
@@ -28,21 +20,15 @@ import {
   ActiveApplicationSummaryAPIOutDTO,
   InstitutionDetailAPIOutDTO,
   InstitutionContactAPIInDTO,
-  InstitutionUserAPIOutDTO,
   SearchInstitutionAPIOutDTO,
   InstitutionBasicAPIOutDTO,
   CreateInstitutionAPIInDTO,
-  UserRoleOptionAPIOutDTO,
   InstitutionLocationAPIInDTO,
   InstitutionLocationPrimaryContactAPIInDTO,
   AESTCreateInstitutionAPIInDTO,
   PrimaryIdentifierAPIOutDTO,
-  InstitutionUserStatusAPIOutDTO,
-  CreateInstitutionUserAPIInDTO,
-  UpdateInstitutionUserAPIInDTO,
   UserPermissionAPIInDTO,
 } from "@/services/http/dto";
-import { addPaginationOptions, addSortOptions } from "@/helpers";
 
 export class InstitutionService {
   // Share Instance
@@ -125,80 +111,6 @@ export class InstitutionService {
     return ApiClient.Institution.allInstitutionLocations(institutionId);
   }
 
-  mapUserRolesAndLocation(
-    response: InstitutionUserAPIOutDTO[],
-  ): InstitutionUserViewModel[] {
-    return response.map((institutionUser) => {
-      const roleArray = institutionUser.authorizations
-        .map((auth) => auth.authType.role ?? "")
-        .filter((institutionUserRole) => institutionUserRole !== "");
-      const role = roleArray.length > 0 ? roleArray.join(" ") : "-";
-      const locationArray = institutionUser.authorizations
-        .map((auth) => auth.location?.name ?? "")
-        .filter((loc) => loc !== "");
-      const userType = institutionUser.authorizations.map(
-        (auth) => auth.authType.type,
-      );
-      const location = userType.includes("admin")
-        ? ["All"]
-        : locationArray.length > 0
-        ? locationArray
-        : [];
-
-      const viewModel: InstitutionUserViewModel = {
-        id: institutionUser.id,
-        displayName: `${institutionUser.user.firstName} ${institutionUser.user.lastName}`,
-        email: institutionUser.user.email,
-        userName: institutionUser.user.userName,
-        userType,
-        role,
-        location,
-        isActive: institutionUser.user.isActive,
-        disableRemove:
-          AuthService.shared.userToken?.userName ===
-          institutionUser.user.userName,
-      };
-
-      return viewModel;
-    });
-  }
-
-  /**
-   * To get the institution user summary.
-   * @param paginationOptions
-   * @param institutionId
-   * @returns All the institution users.
-   */
-  public async institutionUserSummary(
-    paginationOptions: PaginationOptions,
-    institutionId?: number,
-  ): Promise<InstitutionUserAndCountForDataTable> {
-    let url = institutionId
-      ? `institution/${institutionId}/user`
-      : "institution/user";
-    url = addPaginationOptions(
-      url,
-      paginationOptions.page,
-      paginationOptions.pageLimit,
-      "?",
-    );
-    url = addSortOptions(
-      url,
-      paginationOptions.sortField,
-      paginationOptions.sortOrder,
-    );
-
-    if (paginationOptions.searchCriteria) {
-      url = `${url}&${PaginationParams.SearchCriteria}=${paginationOptions.searchCriteria}`;
-    }
-    const response: PaginatedResults<InstitutionUserAPIOutDTO> =
-      await ApiClient.Institution.institutionUserSummary(url);
-    return {
-      results: this.mapUserRolesAndLocation(response.results),
-      count: response.count,
-    };
-  }
-
   /**
    * Create the user authorizations to be associate with the user.
    * @param isAdmin must be created as an admin user.
@@ -242,137 +154,6 @@ export class InstitutionService {
     }
 
     return permissions;
-  }
-
-  /**
-   * Create a user, associate with the institution, and assign the authorizations.
-   * @param bceidUserId user BCeID id from BCeID Web Service (e.g. SomeUserName)
-   * that will have its data retrieved to be created on SIMS.
-   * @param isAdmin must be created as an admin user.
-   * @param isLegalSigningAuthority for admin users, define if the role
-   * legal-signing-authority should be associated with the user.
-   * @param locationAuthorizations for non-admin users, the individual permission
-   * for every location.
-   * @param institutionId institution to have the user associated. If not provided the
-   * token information will be used, if available.
-   */
-  async createInstitutionUserWithAuth(
-    bceidUserId: string,
-    isAdmin: boolean,
-    isLegalSigningAuthority: boolean,
-    locationAuthorizations: LocationAuthorization[],
-    institutionId?: number,
-  ): Promise<void> {
-    const userPayload = { bceidUserId } as CreateInstitutionUserAPIInDTO;
-    userPayload.permissions = this.createUserPermissions(
-      isAdmin,
-      isLegalSigningAuthority,
-      locationAuthorizations,
-    );
-    await ApiClient.Institution.createInstitutionUserWithAuth(
-      userPayload,
-      institutionId,
-    );
-  }
-
-  /**
-   * Updates an existing user authorizations.
-   * @param userName unique user name from SIMS (e.g. someGuid@bceid).
-   * @param isAdmin must be created as an admin user.
-   * @param isLegalSigningAuthority for admin users, define if the role
-   * legal-signing-authority should be associated with the user.
-   * @param locationAuthorizations for non-admin users, the individual permission
-   * for every location.
-   */
-  async updateInstitutionUserWithAuth(
-    userName: string,
-    isAdmin: boolean,
-    isLegalSigningAuthority: boolean,
-    locationAuthorizations: LocationAuthorization[],
-  ): Promise<void> {
-    const userPayload = {} as UpdateInstitutionUserAPIInDTO;
-    userPayload.permissions = this.createUserPermissions(
-      isAdmin,
-      isLegalSigningAuthority,
-      locationAuthorizations,
-    );
-    await ApiClient.Institution.updateInstitutionUserWithAuth(
-      userName,
-      userPayload,
-    );
-  }
-
-  public async getInstitutionLocationUserDetails(
-    userName: string,
-  ): Promise<InstitutionUserAPIOutDTO> {
-    return ApiClient.InstitutionLocation.getInstitutionLocationUserDetails(
-      userName,
-    );
-  }
-
-  public async updateUserStatus(userName: string, userStatus: boolean) {
-    return ApiClient.InstitutionLocation.updateUserStatus(userName, userStatus);
-  }
-
-  public async prepareAddUserPayload(
-    isAdmin: boolean,
-    selectUser: UserRoleOptionAPIOutDTO,
-    adminRole: string,
-    institutionLocationList: InstitutionLocationsDetails[],
-  ) {
-    return {
-      userId: selectUser.code,
-      userType: isAdmin ? "admin" : undefined,
-      userRole: isAdmin ? adminRole : undefined,
-      userGuid: selectUser.id,
-      location: !isAdmin
-        ? (institutionLocationList
-            .map((el: InstitutionUserWithUserType) => {
-              if (el.userType?.code) {
-                return {
-                  locationId: el.id,
-                  userType: el.userType?.code,
-                };
-              }
-            })
-            .filter((el: any) => el) as InstitutionUserRoleLocation[])
-        : undefined,
-    } as InstitutionUserAuthDetails;
-  }
-
-  public async prepareEditUserPayload(
-    institutionUserName: string,
-    isAdmin: boolean,
-    adminRole: string,
-    institutionLocationList: InstitutionUserWithUserType[],
-  ) {
-    return {
-      userGuid: institutionUserName ? institutionUserName : undefined,
-      userType: isAdmin ? "admin" : undefined,
-      userRole: isAdmin ? adminRole : undefined,
-      location: !isAdmin
-        ? (institutionLocationList
-            .map((el: InstitutionUserWithUserType) => {
-              if (el.userType?.code) {
-                return {
-                  locationId: el?.id,
-                  userType: el.userType?.code,
-                };
-              }
-            })
-            .filter((el: any) => el) as InstitutionUserRoleLocation[])
-        : undefined,
-    } as InstitutionUserAuthDetails;
-  }
-
-  public async getMyInstitutionDetails(authHeader?: any) {
-    return ApiClient.Institution.getMyInstitutionDetails(authHeader);
-  }
-
-  public async getMyInstitutionLocationsDetails(authHeader?: any) {
-    return ApiClient.InstitutionLocation.getMyInstitutionLocationsDetails(
-      authHeader,
-    );
   }
 
   public async getLocationsOptionsList(): Promise<OptionItemDto[]> {
@@ -453,12 +234,12 @@ export class InstitutionService {
   }
 
   /**
-   * Get the user status from institution perspective returning the
-   * possible user and institution association.
-   * @returns information to support the institution login process and
-   * the decisions that need happen to complete the process.
+   * Get location details of logged in user.
+   * @returns location details.
    */
-  async getInstitutionUserStatus(): Promise<InstitutionUserStatusAPIOutDTO> {
-    return ApiClient.Institution.getInstitutionUserStatus();
+  async getMyInstitutionLocationsDetails(authHeader?: any) {
+    return ApiClient.InstitutionLocation.getMyInstitutionLocationsDetails(
+      authHeader,
+    );
   }
 }
