+++ sources/packages/web/src/services/http/InstitutionLocation.ts
@@ -4,11 +4,9 @@ import {
   InstitutionLocationFormAPIOutDTO,
   ActiveApplicationDataAPIOutDTO,
   OptionItemAPIOutDTO,
-  InstitutionUserAPIOutDTO,
-  UserActiveStatusAPIInDTO,
-  InstitutionUserLocationsAPIOutDTO,
   InstitutionLocationPrimaryContactAPIInDTO,
   InstitutionLocationAPIInDTO,
+  InstitutionLocationsAPIOutDTO,
 } from "@/services/http/dto";
 
 export class InstitutionLocationApi extends HttpBaseClient {
@@ -43,39 +41,6 @@ export class InstitutionLocationApi extends HttpBaseClient {
     );
   }
 
-  public async getInstitutionLocationUserDetails(
-    userName: string,
-  ): Promise<InstitutionUserAPIOutDTO> {
-    return this.getCallTyped<InstitutionUserAPIOutDTO>(
-      this.addClientRoot(`institution/user/${userName}`),
-    );
-  }
-
-  public async updateUserStatus(
-    userName: string,
-    userStatus: boolean,
-  ): Promise<void> {
-    try {
-      await this.patchCall<UserActiveStatusAPIInDTO>(
-        this.addClientRoot(`institution/user-status/${userName}`),
-        {
-          isActive: userStatus,
-        },
-      );
-    } catch (error: unknown) {
-      this.handleAPICustomError(error);
-    }
-  }
-
-  public async getMyInstitutionLocationsDetails(
-    header?: any,
-  ): Promise<InstitutionUserLocationsAPIOutDTO[]> {
-    return this.getCallTyped<InstitutionUserLocationsAPIOutDTO[]>(
-      this.addClientRoot("institution/my-locations"),
-      header,
-    );
-  }
-
   public async getOptionsList(): Promise<OptionItemAPIOutDTO[]> {
     return this.getCallTyped<OptionItemAPIOutDTO[]>(
       this.addClientRoot("location/options-list"),
@@ -90,6 +55,19 @@ export class InstitutionLocationApi extends HttpBaseClient {
       this.addClientRoot(
         `location/${locationId}/active-application/${applicationId}`,
       ),
+    );
+  }
+
+  /**
+   * Get location details of logged in user.
+   * @returns location details.
+   */
+  async getMyInstitutionLocationsDetails(
+    header?: any,
+  ): Promise<InstitutionLocationsAPIOutDTO[]> {
+    return this.getCallTyped<InstitutionLocationsAPIOutDTO[]>(
+      this.addClientRoot("location"),
+      header,
     );
   }
 }
