+++ sources/packages/web/src/services/http/dto/InstitutionLocation.dto.ts
@@ -85,3 +85,9 @@ export interface ActiveApplicationSummaryAPIOutDTO {
   scholasticStandingId?: number;
   applicationScholasticStandingStatus: ApplicationScholasticStandingStatus;
 }
+
+export interface InstitutionLocationsAPIOutDTO {
+  id: number;
+  name: string;
+  address: AddressAPIOutDTO;
+}
