+++ sources/packages/web/src/services/http/dto/index.ts
@@ -9,3 +9,4 @@ export * from "@/services/http/dto/Student.dto";
 export * from "@/services/http/dto/Report.dto";
 export * from "@/services/http/dto/Pagination.dto";
 export * from "@/services/http/dto/ApplicationException.dto";
+export * from "@/services/http/dto/Institution-user.dto";
